Create task7 (1)
In this example, the http.createServer method is used to create an HTTP server that listens on port 3000. When the server receives a request, the res.end method is used to send a response to the client with the message "Welcome to my Node Js project". The server logs a message to the console when it starts listening on the specified port.