Create 7(2)
In this example, a switch statement is used to determine the response based on the value of req.url. The res.end method is used to send a response to the client with the appropriate message. If the URL does not match any of the defined cases, a default response of "Invalid URL" is returned.

Note that this is a very basic example, and in a real-world application you would likely want to handle more complex logic, return different HTTP status codes based on the request, and potentially use different response headers.